Complete exercises 6.10 - 6.11
6.10 Better anecdotes, step8
"Install Redux Toolkit for the project. Move the Redux store creation into the file store.js and use Redux Toolkit's configureStore to create the store.

Change the definition of the filter reducer and action creators to use the Redux Toolkit's createSlice function."

6.11 Better anecdotes, step9
"Change also the definition of the anecdote reducer and action creators to use the Redux Toolkit's createSlice function."